Refactor code to use object-oriented design patterns

+++ ping.h
@@ -2,17 +2,14 @@
 #define PING_H
 
 #include <string>
-#include <map>
 #include <vector>
+#include <map>
 #include <tuple>
 #include <future>
 #include <chrono>
 #include <sstream>
 #include <iomanip>
-#include <cstdlib>
-#include "db.h"
 
-// 执行ping操作并返回结果
-std::vector<std::tuple<std::string, std::string, bool, long long, std::string>> performPing(const std::map<std::string, std::string>& hosts);
+std::vector<std::tuple<std::string, std::string, bool, short, std::string>> performPing(const std::map<std::string, std::string>& hosts);
 
 #endif // PING_H
